Fix tone passed to sound_connector
In the merging with the changes from the ev3dev-lang-python
repository, the behaviour here got changed. However, it needs to pass
a dictionary to ```self.beep``` instead of an argument list meant for
the linux 'beep' command. This commit reinstates correct behaviour.

Furthermore, we also fix a test where the expected value was a
calculated floating point which was just a last digit off from the one
that is calculated by the actual code.